cli: fix undefined: config.MysqldConfig.
